requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.1.1
django-cors-headers==4.9.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
flake8==5.0.4
flake8-docstrings==1.7.0
fonttools==4.58.5
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
mccabe==0.7.0
numpy==2.3.1
openpyxl==3.1.2
packaging==25.0
pep8-naming==0.13.3
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.11
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pygame==2.5.2
pyparsing==3.2.3
pytest==8.3.5
pytest-rerunfailures==15.1
pytest-timeout==2.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2024.2
requests==2.32.3
six==1.17.0
snowballstemmer==3.0.1
sqlparse==0.5.2
tomli==2.2.1
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.11.0
